Several more bug fixes found through cross-Python-version testing.
Got rid of the template format for the build_optional_ext fail_message altoger.
The Python 3 version in ConfigParser assumes that anything begging with '%'
must be an interpolation variable.  I suppose I could use some other format
like with a $, but it's not worth it.  Now build_optional_ext always outputs a
reasonable default message "optional extension such and such failed to build: 
reason" and if a custom fail_message is defined it is output after the default
one.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci.distutils/trunk@14995 fe389314-cf27-0410-b35b-8c050e845b92