Make sure not to import lib2to3.refactor on Python 2, where it is not needed. This can cause an obscure error, where importing lib2to3.refactor causes the parser generator included in libto3 to read and parse a Grammar.txt file for the Python grammar, and then save in-memory grammar as a pickled object back to lib2to3's package directory (if being run for the first time).
If installing stsci.distutils on a system where 2to3 is installed has never
been run, this can occur during stsci.distutils import.  It's not likely
for lib2to3 to be installed in Python 2. But if it is, it's even less likely
that it's ever been run.  Upon trying to save the grammar pickle this can
result in a SandboxViolation error if installing with easy_install.

Hypothetically this could be a problem on Python 3 systems as well, but it's
still very unlikely, so we'll deal with that if we get there.  That could
require a more involved workaround such as adding lib2to3's package directory
to the DirectorySandbox exceptions.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci.distutils/trunk@27464 fe389314-cf27-0410-b35b-8c050e845b92